about page basic. Can add more as we develop more. Keeping it simple to start.
