+++ client/src/App.js
@@ -2,12 +2,12 @@ import { Route, Routes } from "react-router-dom";
 import './assets/css/App.css';
 
 import Navigation from './components/Navigation';
-import About from './components/About';
-import Contact from './components/Contact';
-import FAQ from './components/FAQ';
-import Home from './components/Home';
-import Packages from './components/Packages';
-import Venues from './components/Venues';
+import About from './pages/About';
+import Contact from './pages/Contact';
+import Home from './pages/Home';
+import Packages from './pages/Packages';
+import Venue from './pages/Venue';
+import Gallery from "./pages/Gallery";
 import { Nav } from "react-bootstrap";
 
 function App() {
@@ -18,11 +18,11 @@ function App() {
       
       <Routes>
         <Route path="/" element={< Home />}></Route>
-        <Route path="/venues" element={< Venues />}></Route>
+        <Route path="/venue" element={< Venue />}></Route>
         <Route path="/packages" element={< Packages />}></Route>
         <Route path="/about" element={< About />}></Route>
         <Route path="/contact" element={< Contact />}></Route>
-        <Route path="/faq" element={< FAQ />}></Route>
+        <Route path="/gallery" element={< Gallery />}></Route>
       </Routes>
     </div>
   );
