--- client/src/components/About.js
@@ -1,5 +1,0 @@
-export default function About() {
-    return (
-       <h1>About</h1> 
-    )
-}
