+++ client/src/components/Navigation.js
@@ -8,11 +8,11 @@ export default function Navigation() {
             <Navbar.Collapse id='navbarScroll'>
                     <Nav className='navWrap'>
                         <NavLink className='navItem' href='/'>Home</NavLink>
-                        <NavLink className='navItem' href='/venues'>Venues</NavLink>
-                        <NavLink className='navItem login' href='/packages'>Packages/Pricing</NavLink>
-                        <NavLink className='navItem' href='/about'>About Us</NavLink>
+                        <NavLink className='navItem' href='/about'>About</NavLink>
+                        <NavLink className='navItem login' href='/gallery'>Gallery</NavLink>
+                        <NavLink className='navItem' href='/venue'>Venue</NavLink>
+                        <NavLink className='navItem login' href='/packages'>Packages</NavLink>
                         <NavLink className='navItem login' href='/contact'>Contact Us</NavLink>
-                        <NavLink className='navItem login' href='/faq'>FAQ</NavLink>
                     </Nav>
             </Navbar.Collapse>
         </Navbar>
