--- client/src/components/Venues.js
@@ -1,5 +1,0 @@
-export default function Venues() {
-    return (
-       <h1>Venues</h1> 
-    )
-}
