+++ client/src/pages/Gallery.js
@@ -1,0 +1,8 @@
+
+
+
+export default function Gallery() {
+    return (
+       <h1>Gallery</h1> 
+    )
+}
