+++ client/src/pages/Home.js
@@ -1,11 +1,70 @@
+import { Container } from "react-bootstrap";
 
-
-
-
-export default function Home(){
+export default function Home() {
     return (
         <>
-        <h1>This is my home page</h1>
+        <div>
+            <image></image>
+            <h1>
+                Welcome to Cacade Mountain ranch
+            </h1>
+        </div>
+        <div>
+            <body>
+            "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
+            </body>
+        </div>
+        <div>
+            <image></image>
+            <Container>
+                <h2>
+                    The Ranch
+                </h2>
+                <body>
+                    Cascade Mountain lodge and Kelly Mountain lodge 50 acres of woods, mountains and lakes. lorem ipsum
+                </body>
+            </Container>
+        </div>
+        <div>
+            <image></image>
+            <Container>
+                <h2>
+                    Weddings
+                </h2>
+                <body>
+                "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+                </body>
+            </Container>
+        </div>
+        <div>
+            <image></image>
+            <Container>
+                <h2>
+                    Events and reunions
+                </h2>
+                <body>
+                "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+                </body>
+            </Container>
+        </div>
+        <div>
+            <image></image>
+            <Container>
+                <h2>
+                    In Closing.....
+                </h2>
+                <body>
+                "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+                </body>
+            </Container>
+        </div>
+        <div>
+            <Container>
+                <button>Facebook</button>
+                <button>Instagram</button>
+                <button>Twitter</button>
+            </Container>
+        </div>
         </>
     )
 }
