+++ client/src/pages/Venue.js
@@ -1,0 +1,5 @@
+export default function Venue() {
+    return (
+       <h1>Venue</h1> 
+    )
+}
